# docs/swagger.yaml
basePath: /api/v1
definitions:
  authhandler.AuthParams:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  errs.ErrResponce:
    properties:
      error:
        type: string
    type: object
  topichandler.NewTopic:
    properties:
      title:
        type: string
    type: object
  topichandler.Topic:
    properties:
      created_by:
        type: string
      date_created:
        type: string
      date_updated:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  userhandler.NewUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  userhandler.User:
    properties:
      date_created:
        type: string
      date_updated:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
host: 127.0.0.1:8001
info:
  contact: {}
  description: This is a sample server
  title: Booklib API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logins user
      parameters:
      - description: User login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authhandler.AuthParams'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errs.ErrResponce'
      summary: Login user
      tags:
      - auth
  /topics:
    get:
      description: Gets topics
      parameters:
      - description: creator's uuid
        in: query
        name: created_by
        type: string
      - description: topic's title
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/topichandler.Topic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ErrResponce'
      summary: Get topics
      tags:
      - topics
    post:
      consumes:
      - application/json
      description: Creates new topic
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create topic request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/topichandler.NewTopic'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/topichandler.Topic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errs.ErrResponce'
      summary: Create topic
      tags:
      - topics
  /topics/{topicID}:
    get:
      description: Gets topic by id
      parameters:
      - description: Topic id
        in: path
        name: topicID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/topichandler.Topic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.ErrResponce'
      summary: Get topic by id
      tags:
      - topics
  /users:
    post:
      consumes:
      - application/json
      description: Creates new user
      parameters:
      - description: Create user request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userhandler.NewUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/userhandler.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errs.ErrResponce'
      summary: Create user
      tags:
      - users
  /users/{userID}:
    get:
      description: Gets user by id
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User uuid
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.ErrResponce'
      summary: Get user by id
      tags:
      - users
  /users/me:
    get:
      description: Gets user from "subject" field in JWT token
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/userhandler.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errs.ErrResponce'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.ErrResponce'
      summary: Get current user
      tags:
      - users
swagger: "2.0"