# internal/handler/topicHandler/handler.go
package topichandler

import (
	"errors"
	"fmt"
	"log/slog"
	"net/http"

	topicbus "github.com/Javlon721/booklib/internal/bus/topicBus"
	"github.com/Javlon721/booklib/internal/handler/errs"
	"github.com/Javlon721/booklib/internal/handler/middleware"
	"github.com/gofiber/fiber/v3"
	"github.com/google/uuid"
)

type Handler struct {
	topicBus *topicbus.Business
	logger   *slog.Logger
}

func NewHandler(topicBus *topicbus.Business, logger *slog.Logger) *Handler {
	return &Handler{
		topicBus: topicBus,
		logger:   logger,
	}
}

// Create
//
// @Summary Create topic
// @Description Creates new topic
// @Tags topics
// @Param Authorization header string true "JWT token"
// @Param request body NewTopic true "Create topic request body"
// @Accept json
// @Produce json
// @Success 201 {object} Topic
// @Failure 400 {object} errs.ErrResponce
// @Failure 401 {object} errs.ErrResponce
// @Failure 403 {object} errs.ErrResponce
// @Failure 409 {object} errs.ErrResponce
// @Router /topics [post]
func (h Handler) Create(c fiber.Ctx) error {
	var nt NewTopic

	if err := c.Bind().Body(&nt); err != nil {
		return errs.New(errs.InvalidArgument, err)
	}

	userID, err := middleware.GetUserID(c.Context())

	if err != nil {
		return errs.New(errs.Internal, err)
	}

	busTopic, err := toBusTopic(nt, userID)

	if err != nil {
		return errs.New(errs.InvalidArgument, fmt.Errorf("cannot parse topic: %w", err))
	}

	topic, err := h.topicBus.Create(c.Context(), busTopic)

	if err != nil {
		if errors.Is(err, topicbus.ErrTopicAlreadyExists) {
			return errs.New(errs.AlreadyExists, err)
		}
		return errs.New(errs.Internal, err)
	}

	return c.Status(http.StatusCreated).JSON(toHandlerTopic(topic))
}

// GetTopicByID
//
// @Summary Get topic by id
// @Description Gets topic by id
// @Tags topics
// @Param topicID path string true "Topic id"
// @Produce json
// @Success 200 {object} Topic
// @Failure 400 {object} errs.ErrResponce
// @Failure 404 {object} errs.ErrResponce
// @Router /topics/{topicID} [get]
func (h Handler) GetTopicByID(c fiber.Ctx) error {
	topicID, err := uuid.Parse(c.Params("topicID"))

	if err != nil {
		return errs.New(errs.InvalidArgument, fmt.Errorf("cannot parse topicID: %w", err))
	}

	topic, err := h.topicBus.GetUserByID(c.Context(), topicID)

	if err != nil {
		if errors.Is(err, topicbus.ErrTopicNotFound) {
			return errs.New(errs.NotFound, err)
		}
		return errs.New(errs.Internal, err)
	}

	return c.JSON(toHandlerTopic(topic))
}

// GetTopicByID
//
// @Summary Get topics
// @Description Gets topics
// @Tags topics
// @Param created_by query string false "creator's uuid"
// @Param title query string false "topic's title"
// @Produce json
// @Success 200 {object} []Topic
// @Failure 400 {object} errs.ErrResponce
// @Router /topics [get]
func (h Handler) Query(c fiber.Ctx) error {
	qp, err := parseQueryParams(c)

	if err != nil {
		return errs.New(errs.InvalidArgument, err)
	}

	h.logger.Info("handler", "qp", qp)

	filter, err := parseFilter(qp)

	if err != nil {
		return errs.New(errs.InvalidArgument, err)
	}

	h.logger.Info("handler", "filter", filter)

	topics, err := h.topicBus.Query(c.Context(), filter)

	if err != nil {
		return errs.New(errs.Internal, err)
	}

	return c.JSON(toHandlerTopics(topics))
}